+++ cpp/bisMatrixTransformation.h
@@ -97,7 +97,7 @@ class bisMatrixTransformation : public bisAbstractTransformation {
   virtual void serializeInPlace(unsigned char* output);
 
   /** returns size needed to serialize this object in bytes */
-  virtual int getRawSize();
+  virtual long getRawSize();
 
 
   // No
